+++ compiler/src/org.graalvm.compiler.phases.common/src/org/graalvm/compiler/phases/common/LoweringPhase.java
@@ -225,6 +225,11 @@ public class LoweringPhase extends BasePhase<PhaseContext> {
         this.loweringStage = loweringStage;
     }
 
+    @Override
+    protected boolean shouldDumpBeforeAtBasicLevel() {
+        return loweringStage == LoweringTool.StandardLoweringStage.HIGH_TIER;
+    }
+
     /**
      * Checks that second lowering of a given graph did not introduce any new nodes.
      *