+++ compiler/src/org.graalvm.compiler.printer/src/org/graalvm/compiler/printer/BinaryGraphPrinter.java
@@ -104,7 +104,7 @@ public class BinaryGraphPrinter implements GraphPrinter {
     private static final int KLASS = 0x00;
     private static final int ENUM_KLASS = 0x01;
 
-    static final int CURRENT_MAJOR_VERSION = 2;
+    static final int CURRENT_MAJOR_VERSION = 1;
     static final int CURRENT_MINOR_VERSION = 0;
 
     static final byte[] MAGIC_BYTES = {'B', 'I', 'G', 'V'};
@@ -173,14 +173,19 @@ public class BinaryGraphPrinter implements GraphPrinter {
         return snippetReflection;
     }
 
+    @SuppressWarnings("all")
     @Override
     public void print(Graph graph, Map<Object, Object> properties, int id, String format, Object... args) throws IOException {
         writeByte(BEGIN_GRAPH);
-        writeInt(id);
-        writeString(format);
-        writeInt(args.length);
-        for (Object a : args) {
-            writePoolObject(a);
+        if (CURRENT_MAJOR_VERSION >= 3) {
+            writeInt(id);
+            writeString(format);
+            writeInt(args.length);
+            for (Object a : args) {
+                writePropertyObject(a);
+            }
+        } else {
+            writePoolObject(formatTitle(id, format, args));
         }
         writeGraph(graph, properties);
         flush();
@@ -352,6 +357,7 @@ public class BinaryGraphPrinter implements GraphPrinter {
         return getClassName(klass.getComponentType()) + "[]";
     }
 
+    @SuppressWarnings("all")
     private void addPoolEntry(Object object) throws IOException {
         char index = constantPool.add(object);
         writeByte(POOL_NEW);
@@ -382,8 +388,14 @@ public class BinaryGraphPrinter implements GraphPrinter {
         } else if (object instanceof NodeClass) {
             NodeClass<?> nodeClass = (NodeClass<?>) object;
             writeByte(POOL_NODE_CLASS);
-            writePoolObject(nodeClass.getJavaClass());
-            writeString(nodeClass.getNameTemplate());
+            if (CURRENT_MAJOR_VERSION >= 3) {
+                writePoolObject(nodeClass.getJavaClass());
+                writeString(nodeClass.getNameTemplate());
+            } else {
+                writeString(nodeClass.getJavaClass().getSimpleName());
+                String nameTemplate = nodeClass.getNameTemplate();
+                writeString(nameTemplate.isEmpty() ? nodeClass.shortName() : nameTemplate);
+            }
             writeEdgesInfo(nodeClass, Inputs);
             writeEdgesInfo(nodeClass, Successors);
         } else if (object instanceof ResolvedJavaMethod || object instanceof Bytecode) {